Game changes
Changed the image used for iron mines and their blueprint.  Made it so building select list has a standard size. Made a connection in iron mine selectable. Added code in build manager so it works with building select list.
Added a function in selectable building to get the building data associated with the specific selectable. Added the image for the steel mine. Ran all these things in game to make sure they worked together