Added asset, wood cutter, moved files
Added a log cabin image. Set the log cabin image to be the image for the wood cutter scene and the blueprint of the wood cutter. Moved all selectable scenes to a new folder called selectables. Created a selectable for the wood cutter and added it to the building select list.