Work done for currency panel
Added a working animation to the currency panel for the rotating coin. Removed the building cost field from selectable buildings and added them to the building data base class. Added visitor pattern in the BuildingInfo class working together with Building data class and its inheritors. Split parts of a spritesheet into individual images to use in the animation. Created a sprite frames resource from a sprite atlas. Set the building cost of some of the buildings resources. Set mouse filter stop for most of the UI scenes