BuildingInfo panel text formating
After the merge from main, added specific text formating in the buildinginfo panel to account for area gathering buildings. Also added a pollutions text formatter. Made it so that more complex buildings reuse the text formating of the less complex buildings. Added a unique version of the accept function to areagathering building data