Temporary sell button, scripts
Made a temporary no styled button for selling production. Added parts in the script in BuildingInfo for disabling and enabling the sell production button depending on the building type. Added a current building variable in the BuildingInfo script to keep track of the currently targeted building. Added code in the scripts for both gathering building and production building since they have different functions for producing goods.