storage for byproducts, filepath for biomass image
Added function for hiding the buttons in a stored resource panel. Added a filepath for the biomass image in enums. Added specific scenarios in building info panel for when a resource is a byproduct. Added a comment in accept function of building data explaining the way to use the function. Removed upkeep from road.