Bugfixes for merge
Fixed a bug where resources were not accurately removed from warehouses when sold through the buildinginfo menu. Fixed a bug where the selling tab in the building info panel could be accessed on a warehouse by first pressing a different building, selecting the selling tab and then pressing the warehouse. This would then lead to a crash if the player pressed a selling mode.