Mode updates for buildings
Made some changes so that the building info tab showing the mode of the building selected now shows 'Mixed' when selecteing multiple buildings with different modes. Added the ProductionBuildingMode enum to the Enums script. Added text interpretations of the modes to Enums as well as a function for getting these interpretations. Made it so that production buildings no longer use a bool for checking what mode they are in but instead use the ProductionBuildingMode enum.